requirements.txt
discord.py
PyNaCl
gTTS